리듀서 초기값을 공통으로 사용하도록 fixtures 로 이동
